parser now takes input as a set of files and preamble

+++ src/mos/include/file.h
@@ -2,8 +2,9 @@
 #define MOS_FILE_H
 
 #include "alloc.h"
+#include "types.h"
 
-void        file_read(allocator *, char const *, char **, size_t *);
+void        file_read(allocator *, char const *, char **, u32 *);
 char const *file_basename(char const *);
 
 #endif
